+++ packages/class-core/CHANGELOG.md
@@ -22,8 +22,17 @@
 * Change all occurrences of `LionWebJsonChunk` type in delta serialization types to `LionWebJsonDeltaChunk`.
 * Expose a `serializeAsDeltaChunk` function — and propagate its use to the delta protocol packages.
   * Fix a bug in the value managers for multi-valued containments:
-    a call to `insertAtIndex` with a child that’s already in the managed multi-valued containment now results in a `ChildMovedInSameContainment` delta with a correct `indexOffset`
+    a call to `insertAtIndex` with a child that’s already in the managed multi-valued containment now results in a `ChildMovedInSameContainmentInSameParent` delta with a correct `indexOffset`
     (instead of the effective `newIndex`).
+* Perform **breaking** renamings of types:
+  * `ChildMovedFromOtherContainmentDelta` &rarr; `ChildMovedFromContainmentInOtherParentDelta`
+  * `ChildMovedFromOtherContainmentSerializedDelta` &rarr; `ChildMovedFromContainmentInOtherParentSerializedDelta`
+  * `ChildMovedAndReplacedFromOtherContainmentDelta` &rarr; `ChildMovedAndReplacedFromContainmentInOtherParentDelta`
+  * `ChildMovedAndReplacedFromOtherContainmentSerializedDelta` &rarr; `ChildMovedAndReplacedFromContainmentInOtherParentSerializedDelta`
+  * `ChildMovedInSameContainmentDelta` &rarr; `ChildMovedInSameContainmentInSameParentDelta`
+  * `ChildMovedInSameContainmentSerializedDelta` &rarr; `ChildMovedInSameContainmentInSameParentSerializedDelta`
+  * `ChildMovedAndReplacedInSameContainmentDelta` &rarr; `ChildMovedAndReplacedInSameContainmentInSameParentDelta`
+  * `ChildMovedAndReplacedInSameContainmentSerializedDelta` &rarr; `ChildMovedAndReplacedInSameContainmentInSameParentSerializedDelta`
 
 
 ## 0.9.2